fix: 修复底层库多项代码缺陷
1. ds1302.c: 修复 SCK 未定义变量错误 (SCK→SCL)，使代码可正常编译
2. ds1302.c: 修复 datetime_write/datetime_read 函数签名与头文件不一致，改为接受参数
3. ds1302.c: 移除未使用的 time_init 全局变量
4. onewire.h: 修复 read_temperature 返回类型 (float→uint8_t) 与实现一致
5. iic.c: 修复 Read_2k 缺少设备地址字节 (0xA0)，补充 I2C 起始后的设备寻址
6. .gitignore: 添加 .trae-html-share-packages/ 忽略规则

+++ 底层库/onewire.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-float read_temperature();
+uint8_t read_temperature();
 
 #endif
